Improve error messaging around auth failures
Rather than just dumping the raw HTTPError exception, we should provide
some additional guidance for the end user in the specific case that it's
still returning 401 when we attempt to authenticate. Because we retry
authentication to reissue tokens, if the authentication ultimately fails
with 401 then it points to bad credentials or the container/registry is
not actually available. Some registries return 401 instead of a 404.

Change-Id: Ic8da92d98656bcea7db5701b44eeadc1d5c87198
Closes-Bug: #1860040