Always convert to docker media format
Previously if the container config mediaType was unset or set to
something other than a docker format, we would just pass that on. This
caused problem with stricter checking of the metadata performed by some
versions of Podman. Since we're always converting the layer metadata
information to docker, we should ensure that the config media type is
properly set to a valid docker distribution format. This change ensure
that when we push the image config, that the mediaType is set to match
the contents type.

Change-Id: I6b6a8231f6881fdb53084a437ee24fd852ebdd55
Closes-Bug: #186585
(cherry picked from commit e7d1e59283e13da0d6a253df0a29a243444ed58e)