Remove workbook related to plan environment
It seems the workbook for updating plan environment was only used
with tripleo-ui. Now that tripleo-ui is gone, we can safely remove
it.

It also removes the two mistral action entry_points, but left the
action plugin there as it's being called from container_images.py.
We'll clean that in a later patch.

Story: 2007212
Task: 38577
Task: 38578
Change-Id: I2b72ad74455bc5d6cb4c6a426c831d55a0545179