tests: Get RoleData correctly
In Ie543782569de14d56bc41740611f7512e8357a22 a new entry was added
into the fake stack, but the indexes for any test accessing the
FAKE_STACK were not changed, effectively breaking several tests
silently.

This patch removes the direct referencing of indexes in FAKE_STACK
and uses the config utilities to get the data correctly. It also
corrects the reference to the constant UPGRADE_STEPS_MAX which no
longer exists thanks to Ie03084bb599b7b06aeeb321d2a7938a908487788

We also add mock_assert_called to several of the tests to ensure
that if the fake_role list is empty, the test will fail.

Change-Id: I08757ab389f35de52814fba48722e9325e50ab21