[Queens-only] make sure cellv2 host discovery task only run once
The discovery task is correct triggered via delegate only on a single node,
but all computes delegate the job to this single host. It is possible that
this rund multiple times in parallel and fails as nova-manage commands are
not meant to run multiple times at the same time.

This makes sure the task only gets triggered once.

Change-Id: I072baf357836c55ad37257158b83f25c16b7f46f
Closes-Bug: #1862321