[Train-Only] Expose --limit, --skip-tags, and --tags in config-download.
The Mistral workflows can now consume these ansible-playbook arguments
and using them during config-download (during the deploy and update
only). Separate arguments align with the new Ansible Runner which cannot
accept any arbitrary or random argument.

This is only for Train because in Ussuri we no longer run Ansible
through Mistral. In Ussuri, all of the Ansible Runner code is in place
to allow the usage of these additional flags.

Change-Id: I2a8bcd6f8e365058d667fccb3d47625ab307b9e4
Signed-off-by: Luke Short <ekultails@gmail.com>