Improve step task check to cater for whitespace
Currently a conditional of '(step | int) == 1' will not pass
the regex check because of the additional whitespace on either
side of the filter. There are many other whitespace variations
which could also make it fail, so rather than add a bunch of
\s* into the regex, we opt to remove all whitespace from the
string instead before doing the regex match.

This ensures that we're only running the step tasks when we
should, rather than having these tasks added into all generated
plays.

Change-Id: I589406f30f7b08bc28d05632bd97392f8b3623bf