Don't allow provisioned in the defaults section
The following baremetal provisioning yaml will cause the expand roles
action to go into an infinite loop

- name: Controller
  count: 3
  defaults:
    provisioned: false

This is because instance entries are appended using default values
until the desired number of provisioned instances is achieved, which
never happens because provisioned is false.

This change prevents this condition by adding validation to never
allow provisioned to be specified in the defaults section.

Change-Id: Id0abff98071cdb556c22c0ba42bca0343a2c30ad
Closes-Bug: #1864557