Modify roles to remove unused services
Historically we've included unused services in all the roles as a way
for users to enable the services for a given role. By default many are
set to OS::Heat::None. For large scale deployments this results in many
extra resources managed by heat.

This change adds a function to look at the resource registry of the
environment and will remove any unused services from the roles data.
Removing these services from the roles data reduces number of heat
resource and hence possibly little deployment time.

Change-Id: I54539f7dd694e0804006470221e55fd3cafea539
Co-Authored-By: Alex Schultz <aschultz@redhat.com>
Co-Authored-By: Thomas Herve <therve@redhat.com>