Use config for layers
Podman keeps a copy of the metadata around in the local container
storage however this may not actually represent the actual layers being
used in the storage. Instead of using the manifest file as the source of
truth for the container layers, we need to look into the container image
config for the layer ids for the local storage. This change adds a
function to swap out the layer metadata with layer information from the
image config if available. If ultimately we cannot find the layers we
throw an exception indicating that it is not found.

Closes-Bug: #1864660
Change-Id: Iec115bfa1ef00112705785a5b3e6ed7dc222070e
(cherry picked from commit 2950738526d7b93553102ceb77717d31c445caa2)