Remove invalid mistral envocation and references
This change removes the use of a mistral workflow from our deployment process
which assertains the ansible deployment status. This is being removed because
deployments are not longer executed via mistral so any data provided by mistral
would be wrong.

This change also removes the mistral workflow reference in our documentation
strings so it does not confuse folks maintaining the code.

Story: 2007212
Task: 38962
Task: 38961

Change-Id: I94fed2dd4968879d0ac3ea775dd498de20b78130
Signed-off-by: Kevin Carter <kecarter@redhat.com>