Add libcgroup-tools to libvirt container
To reuse the tripleo containers when running libvirtd container
on OCP we need to run it via a wrapper script and create our own
cgroups and run libvirtd using cgexec to break out the default
OCP cgroup of the pod. Otherwise if the libvirtd pod gets
deleted/recreated all VMs get shutdown as OCP cleans up the
resources owned by the pod.

Change-Id: I469b4f51c57e1651f42064b0134aa3916af72d80