Fix multistack inventory
The static inventory merging was setting the role hosts to the _meta hostvars.
This is a special key to set hostvars for dynamic inventories. It contains all
hosts in the inventory to prevent the dynamic inventory script being called
with --host for each host.

It was not used by the single stack inventory when generating a static
inventory. Instead an evil magic flag (host_format_dict), switched the
behaviour to embed the hostvars.

This flag is now removed and replaced with a param to the list method and
calling the write_static_inventory methods no longer magically mutates
the object. The logic for static vs dynamic inventory should be clearer.

The interface for the multi stack inventory has also change to be more
consistent with the single stack interface.

Unit tests now cover both static and dynamic multistack inventories.

Change-Id: Ife14dbe04dd11db44a944f98373f63e01dfbb8d8
Depends-On: If022a4647aaebc4b212bba188cc959ff2f8f9311
Closes-bug: #1849073