Refactor registry request actions
In order to handle re-authentication at request time rather than
letting tenancity trigger it, this change creates a session helper class
that wraps the various get/patch/push/put actions that we use in the
image uploader to perform a single re-autentication action when we get a
401 from a registry.  Previously we would let the reauthentication occur
when tenancity would retry various functions.  There are several
functions which perform muiltple requests that could exceed the TTL for
authentication tokens. This leads to failures that could have been
prevented if we re-auth at the time of the request rather than retrying
the entire action.

Change-Id: I2fc2072f0ab8b761a2b8079f746c3b91cacc8733
Related-Bug: #1867981