Set scp_if_ssh to True in ansible.cfg
To avoid the warning "sftp transfer mechanism failed" during Ansible
run (also reported here:
https://github.com/ansible/ansible/issues/22127); let's set scp_if_ssh
to True.

The default is False; which means it'll first try sftp (which fails
since we use pipeline) and fallback on scp. To avoid the warning, let's
set it to True directly, so we use scp.

Change-Id: I637d850d0de75e60835d229b36b85e699b037543
Closes-Bug: #1869735