Preserve existing placement password during extraction
nova-placement previously used NovaPassword. When it was split out into
the standalone placement service a new password param was added.

This patch ensures that when the extraction occurs, PlacementPassword
defaults to the existing NovaPassword value so that we do not have to
update the placement password in keystone.
Works around the fact that the upgrade playbook currently does not run
the required tasks to update keystone resources.

Closes-bug: #1870528
Change-Id: I9c84a48f56ac372ad612e2cf68533e256117eff1
(cherry picked from commit 2698da0ccbc59a0e8ab95fd36a391c554d6d057d)