Allow ansible_group_vars to be overriden by RoleGroupVars
We have introduced {{role.name}}ExtraGroupVars in THT; which allows an
operator to override any Ansible group vars.

In this patch, we will override the ansible_group_vars (from RoleData)
with the RoleGroupVars defined by THT.

It also updates the unit tests to cover 2 scenarios:

- chrony_host set to localhost in RoleData but overriden by 192.168.2.1
  in RoleGroupVars; in actual ansible group vars we end up with
  192.168.2.1.
- chrony_acl set to none in RoleData, never overriden; in actual ansible
  group vars we end up with none.
- chrone_foo isn't set in Role Data, but only by RoleGroupVars and it
  ends up being in actual ansible group vars.

Also, it removes the "if role_name not in role_group_vars" condition, as
the condition is never met. There is always a role_name key in
role_group_vars (e.g. max_fail_percentage, etc).

Change-Id: I2684174bcd9744f0ec0c1ec42cf11d2c467fc7ba