[train-only] Update tripleo-ansible path for fernet playbook
The rotate-keys.yaml formerly lived in tripleo-common, but was later
refactored into tripleo-ansible, which makes sense because it's an
ansible playbook. When that happened, tripleo-common was updated to
point to the location of the new playbook [0].

However, it appears that the playbooks are installed into
/usr/share/ansible/tripleo-playbooks/ and not
/usr/share/tripleo-ansible/playbooks. This will cause errors invoking
key rotation since mistral can't find the playbook to execute:

  ERROR! A playbook must be a list of plays, got a <class
  'ansible.parsing.yaml.objects.AnsibleUnicode'> instead

This commit fixes the issue by updating the path to point to the
installed location of rotate-keys.yaml. This fix is only applicable to
stable/train since mistral was removed from master.

[0] https://opendev.org/openstack/tripleo-common/commit/50e464c6f47f1f70d00ec4dc5b22447f6a3846f6

Change-Id: I74b64476b74344eb7d4cf1c0635f07a5fc128d3c