[stable/train only] Use nodes as hostnames if ironic lookup fails
This allows us to use both instance ids and hostnames with
node delete when using mistral.

This also ensures that we fail early when ansible fails to
match host pattern and ignores a wrong node provided.

Note: This won't work when a mix of hostnames and instance ids
are used with node delete command.

Change-Id: I23fc92dfdb1808dec59f0ff7f96910ab8b5ccf45
Closes-Bug: #1871155