[Train only] Fix misused timeout values in the deployment workflow
Mistakingly and historically
tripleo.deployment.v1.config_download_deploy used to take wrong value
for its config_download_timeout input. That input defines global
deployment timeout but the caller passes ssh connection timeout for it.

This is fixed in the scope of the related bug (the caller on the
tripleoclient side).

Although to pertain the established tradition for setting connection
timeout to some timeout, add missing ssh connection timeout input and
make it matching the passed run_ansible's timeout value.

This needs to be fixes for Train and earlier releases only (the
workflow is no longer used after Train)

Change-Id: Ia0c92af6cdcdd66dd5eabb945b4398e9228be464
Related-Bug: #1868063
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>