Use a full string match for container images
When building the list of container images to pass to kolla-build, we
currently use the plain image names. However, since these are regexes,
we can hit problems with substring matches. For example, the 'rabbitmq'
image matches 'rabbitmq-3.7.10', an image which is being added to kolla
in the stable/train branch that Tripleo is not interested in.

This change uses a full string match to avoid this issue.

Change-Id: Ic78df6a98b7f0c796ec9e112f572c9b50cba396b