utils/config: allows a when to have boolean conditions
In Ansible, a condition in when doesn't have to be a string, and can in
theory be a boolean. In that case, we don't want to use regex on it, we
would get a TypeError.

Let's just continue the loop and move on.

Change-Id: Ie00144d821c920973df4495997c0ce376b7d916b
(cherry picked from commit 90dce1dd886762af1ebdebe9ab0b5ba7e810ddd6)