[train-only] Ensure overcloud inventory group exists
In ussuri the "allovercloud" group instead of the "overcloud" group
refers to all overcloud nodes to avoid conflicting with the inventory
group for the "overcloud" stack.
For train only we need to ensure an "overcloud" group exists for any
existing plays that expect this group to refer to all overcloud hosts.

Change-Id: I166647afbf2f893b8ee1981dfb01be336c8f9971
Related-bug: #1875429