Revert "[Train only] Add ssh timeout for baremetal_deploy"
This change breaks node provision because the deploy_roles workflow is
missing a connection_timeout input which is referenced on line #236.

But also I don't understand the intent of adding a connection_timeout
input to these workflows since it is not passed to any of the actions,
and none of the actions call ansible or ssh to remote hosts anyway.

This reverts commit 3d3afa62dc392236dd3191956ed2bf2f05f3b0e1.

Change-Id: I27407eef41d945e3cb23c5305e133c0c5fe5aa12