tcib: move cradox to gnocchi-base & enable ci on scenario001
gnocchi-db-sync requires cradox to be installed if RBD backend is used.
Until we fix it properly in the distgit, let's make sure cradox is in
the base image for gnocchi. It also requires httpd and deps for the
dv_sync to operate when Ceph is installed.

With that patch, we can now test container builds on scenario001.

Change-Id: Id13028ab911ba91d67787ca8247effc02017fcbc