Add interface-names to Rhel8 images
With EL8, net.ifnames is set to 0 in the grub config. This is a change
in the default image which can have user impacts, Details in bug[1].

We already added "interface-names" to centos8 images[2]. With this
change we are adding "interface-names" to remove net.ifnames=0 from
rhel8 image as well.

[1] https://bugs.launchpad.net/tripleo/+bug/1841441
[2] https://review.opendev.org/#/c/712571/

Change-Id: I91567185464997eb77217b14e7efb97cbecf6e77