Set undercloud ansible_connection to local
The Ansible undercloud connection defaults to local but
the inventory object constructor overrode it to SSH for
Mistral. Now that Mistral has been removed we can remove
the override and use local. This also avoids permission
conflicts between the default ansible_ssh_user and the
user running the playbook as per the bug this patch
closes.

Change-Id: Iac7c15572f83e77241bbc1469d3e3debc8ac011b
Closes-Bug:  #1884123