Don't assume default tag exists in container repo
When no tag is set for an entry in ContainerImagePrepare, the default
tag from container-images/container_image_prepare_defaults.yaml will be
assumed, which is typically current-tripleo, or a release version such
as 16.0.

In most cases, this tag will exist. However, when using a satellite with
a content view that has been filtered on tags, it likely won't exist
since content views are often used with container image versions that
are not the latest.

In the case where no tag is set in the entry in ContainerImagePrepare,
and the default tag does not exist in the container repo, the latest tag
from the repo will be assumed instead.

Change-Id: I985ef22c340c4071866c8c51bf303a6f4ee7713c
Closes-Bug: #1886547
Signed-off-by: James Slagle <jslagle@redhat.com>