Add a strict option to per role tasks file generation.
For upgrade and deployment we generate tasks files based on the step
tasks must run in.

By default, all tasks that doesn't have a matching conditional are
included in all step files.

We add a option to be able to control this behavior.

PER_STEP_TASKS becomes a dictionary with key being the current config
collected and the value is an array.  Each element of the array map to
the steps: element 0 is step0 and so on.  That element tell the
_write_tasks_per_step function if we include tasks with no
conditional when it is False or excluded when it is True.

This structure opens four possibilities:

 - allow to have step task of different length: for instance if we
   want to expand post update tasks which have a length of four (not
   DEFAULT_STEPS_MAX)

 - In testing, we can identify the tasks with no conditional by
   putting all normal step to strict (set True for every step) and add
   a extra one set to False.  The extra one will have the steps with
   no conditional.

 - Identify when the conditional matcher in _write_tasks_per_step
   fails.

 - Eventually, set everything to strict to force the developer to
   properly add conditional to every tasks.

As it is, the change has no impact and keep the current default
behavior.

Change-Id: I2033efd47c09707797a4b48a853517d42f584e76
(cherry picked from commit 7c67777f4a256bfc8c5285294769671f0922d479)