Update container image sources in U
All of the container image sources in U were pointed at a respo in
quay.io which does not exist. Additionally the repo name was
incorrect meaning that should it exist it would pull from the wrong
container set. This change updates the repos URL and name to point
to a valid URL.

Change-Id: I8122178601fd8e28bcf0a878c410d348e6539dd6
Signed-off-by: Kevin Carter <kecarter@redhat.com>