[TRAIN-and-older] Reduce wait time to send ansible output
Currently we wait 30 seconds before sending output back to the end user.
This can lead to really large messages and artificial delay for the user
watching the actions.  This change reduces the time we wait to 5 seconds
to ensure the end user gets more frequent updates on the deployment
status.

We dropped this in Ussuri so it only applies to Train and older.

Change-Id: I2f3a08fa543642d2f9d5b5fbe11016b3cfa9f910