Switch 55-heat-config to async
If while a task is executing the ssh connection is severed, ansible will
automagically rerun the command under the covers. This causes problems
for long running 55-heat-config tasks as first process may have written
out the deployed json but not the notify.json that we use use to
determine if it was successful or not. This can lead to a failure
because the process either never runs to completion.  This change
switches the execution to always be run async to ensure that ssh
interruptions won't cause inconsistent failures.

We previously saw a similar issue when invoking the NetworkDeployments
using this process. We've moved the network configurations to the
NetworkConfig task in THT/common/deploy-steps.j2 but this code is still
used to invoked with OS::Heat::SoftwareDeploymentGroup

Change-Id: Ic911bb6d999caf2dc4afd4cff3d44047c03dc8e4
Related-Bug: #1792343
Closes-Bug: #1887846
(cherry picked from commit 43aaaaa1190d7f5e808c906e82abcf116e0508c9)