Stop doing tag_from_label lookup with tag
When using specifying ContainerImagePrepare if a tag is explicitly
provided in a set, the tag_from_label functionality will not be run as
we use the defined tag for the containers. Previously we would still
attempt tag lookups even if we wanted a specific tag.

This would also cause failures during the deployment if using a
namespace that didn't have the defined tag_from_label format but we
defined a specific tag we wanted to use. e.g. using a tag with a md5 vs
a tag_from_label with numbers in it.

Change-Id: I4966641aed1a21be60d915ea58dda78b80fe0e1f
Partial-Bug: #1889122
(cherry picked from commit 014d985f0836340b03d3c2cdeaae8717cfd1ea6b)