Change default DHCP timeout for ironic-python-agent
The ironic-python-agent relies upon obtaining an IP address
through DHCP from one of the connected interfaces. Due to
the design of iPXE and switch configurations, we need the
timeout to be longer for acquisition of an IP address from
DHCP as various switch hold-down timers can interpret the
change from an iPXE environment to a running ramdisk OS
as an event to trigger hold-downs.

Generally this is not an issue once the image is written
to disk, because the port only comes online once, where as
the port effectively bounces and iPXE negotiated behavior
can change as the ramdisk is coming up.

It must be noted that the ironic-python-agent-builder
does default this timeout to 60 seconds, but as TripleO
builds its own images without the builder, this value must
be asserted as well.

Change-Id: I1e9b952d87c0298f568f7c841532a770e25e74a6