Remove check for ./puppet/services when filtering services
It is no longer a way to indentify if a service is containerized
or not. If the service exists in overcloud_containers.yaml.j2,
it's assumed to be containerized.

Also flips logic to not loop through all services in
resource_regstry.

Change-Id: I25fcf887044fdf4d05d5d4cfa268aab6aebdf244
(cherry picked from commit fbb6b57a272cff26b03d8126a20220e6ca509784)