Make build_timeout configurable
Per comment in https://review.opendev.org/#/c/678907/ it would be nice
to have build_timeout configurable by the CLI. And in fact the
job which builds RDO containers for ppc64le needs quite a large timeout,
so that's a good reason to make it configurable now. This is the
necessary stuff to make a python-tripleoclient feasible.

Change-Id: I22387080e69b4881a13d0b37c7c25a75b7740a29