[TRAIN-Only] Dynamically add centos8 namespace
The container namespace is different for the python3/centos8 version of
tripleo. We need to dynamically add it as the default if not explicitly
specified via a ContainerImagePrepare setting. This change adds an
update to the defaults to ensure that the centos8 namespace is changed
in the defaults only when running under python3 and the default is
docker.io/tripleotrain.  This should prevent accidental changes when a
user specifies a specific namespace or the defaults are changed.

Change-Id: I3ff5dc84b15a81d6afcc70ab5dde2d327b3c4714
Related-Bug: #1887176