Wait for lock release during power state change
When power state change is slow, the subsequent provide workflow can
fail because getting a lock on the node times out.

Apparently a node can remain locked for some time after a power state
change, so this issue should be solved by *also* waiting for the node
to be unlocked in the wait_for_power_state action.

Change-Id: I26f23330c50ccf7cb11fb9171d0a82279a497d22
Closes-Bug: #1846791