Add modify_only_with_source
Currently the modify_only_with_labels has to query the container
registry to see what containers we should modify. We can use the source
in from our service to container mapping to limit which containers we
want to update. In the upstream CI, we only want to modify the kolla or
tripleo containers as we want to prevent ceph and other related
containers from being updated.

Change-Id: I4bff2b96f7b13bde808f929c3567dcf167f1eacd
Related-Bug: #1889122