[squash] Backport RHEL modules into base image
This is an unclean backport of RHEL modules into base images.
We include the revert of virt:8:3 change in the images and re-included
it in the base image.

1) Revert "Switch to use virt:8.3 for non-kolla container builds"

This reverts commit 082d3bed48911cafcea56c020c01ba561bd913b4.

2) TCIB: move RHEL modules into base image

Moving the RHEL modules management into the base image will make it
so we will have shared and consistent modules across all container
images; and also configurable from a central place, the base.yaml.

Also, allow to find the module versions from the
{{ tcib_rhel_modules }} variable so later in tripleoclient we can
override the modules versions from the CLI with a new --rhel-modules
argument (which will take comma separated list of modules and their
versions).

Change-Id: I782cbfd3c88b30c801961c6d8ba875d36f59bf42
(cherry picked from commit 640b76608d8194ccbecc77429bca711f37809400)