Fix a couple of Swift healthchecks
Depending on the used procps version, the output of "ps -e" is width
limited and thus not matching the quite long process names. Using
"ps -ef" fixes this, along with a small hack in the piped grep to filter
out grep itself (pgrep itself is not working for this).

Also merged server and replicator healthcheck scripts into single
scripts. Kolla doesn't use separate templates for these images, thus the
replicator images include the server scripts. This has been fixed using
conditional checks on the running processes in the container. Same for
the rsync container, which uses the obect image as well.

Change-Id: I58124fec226d5e765f752d306b411162666d4817