Support basic auth for image registry
If you stand up a docker-distribution registry, basic auth can be
configured with tls to lock down the registry. Currently the only auth
method that is supported is the bearer tokens used by the public
registries.  This change checks the www-authentication header to see if
we should try basic auth or the bearer token auth.

Change-Id: I57599ab3cd8773ae3312930145b9e84244940f41
Closes-Bug: #1893826