Fix handling of default_tag
The previous fix, 589ac8ac0d5a14ce4aa831892d3b0a4a31e29674 was not fully
working as intended because during the templating of the container image
prepare entry, the setting for default_tag was being lost.

This fix simplifies the handling by just checking the initial
mapping_args for the presence of a tag key to determine if the
default_tag flag should be set.

Change-Id: Idfebf62e0e10b709367cfcb6977cf18550a4a4b0
Partial-Bug: #1886547
Signed-off-by: James Slagle <jslagle@redhat.com>