Update container build to better reign in threads
This restructures the buildah build process to ensure we're
managing our thread pool correctly and not spawing more
threads than needed or expected. By updating the thread
process we'll ensure we're building each container only
once and that we're maintaining build order.

Tests have been added to ensure the new methods are called,
and that the data processing is consistent.

Closes-Bug: #1894856
Change-Id: I647212442c378aabbd1887a0b499c19bbc74b5f7
Signed-off-by: Kevin Carter <kecarter@redhat.com>