config-download: allow get_overcloud_config() to run without Swift
Allow the get_overcloud_config() function to work without Swift. It'll
just generate the config into a work directory, and we can optionally
(disabled by default) preserve the config on the host.

This is done so in tripleo_config_download Ansible module we can use
that method to generate the config.

Change-Id: I6c306396cadeb7bcf543b0bbb734f3672f94e97b