Change qemu user id to match previous releases
The qemu user on the host gets created using uid/gid 107. Certificates
on the host, but also the vhost-user sockets created by ovs use this
uid/gid. With the move to TCIB images the default kolla id were
reverted and the previous overwrite dropped. This make e.g. the qemu
processes to fail to use the libvirt-vnc bind mounted certificates.
This change brings back the previous overwrite of the qemu user
uid/gid.

Closes-Bug: #1903508
Related-Bug: #1900986

Change-Id: I54b9d9f341b521b415a6dccc6c78ae7a77821f6f