TCIB: Fix debug mode for oslo process execute
When pushing with buildah, there is no logfile options.
It dumps results of work to stderr.
In debug mode, we want those results logged. But we do not
reconfigure oslo.log defaults, so it skips all details as
it goes with INFO config instead.

Capture debug results of process execution utility for buildah
push and use the current logging context as well.

Also oslo process utils will log the time it took to do
buildah bud builds as a nice free bonus.

Change-Id: I5177408258b6d33daa26239e7b5c4f017b9372cf
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>