Fix localization for horizon container
There's a macro that removes localization data from containers in
order to save space. In order to install localized web apps it
has to be moved the app installed and then moved back

also the packages were being installed twice I also fixed that

It was being done only for openstack-dashboard before now I changed to do
that for all UI web apps.

Closes-Bug: 1905250

/1905250

Change-Id: I66cb4fa785a28cc0a30711ffe834e210c11c7f89