[train-only] Prevent circular inventory group when overcloud stack does not exist
If the overcloud stack does not exist but is in a list of plans and is
the first on that list it can result in the group:

overcloud:
  children:
    overcloud: {}

Requires I98dbbaaea25a0d61674e989eca8d2daf6090ebc4 which omits
non-existing stacks from the plan list.

This patch changes the child group to the first plan that exists or
omits the overcloud group completely if no plans exist.

Change-Id: I60481cf09524dcb8309ee9ff097a4c8bc5dcc12c