Re-factor inventory.py use setdefault pervasively
This refactors the inventory code to pervasively
use 'setdefault' to assign values to all inventory
keys.

The idea is to use multiple sources in order, and
build up the inventory enriching it with more data as
it becomes available.

With the pervasive 'setdefault' usage all values in
the inventory will reflect the value first discovered.

The follow up change
  I9e1851aeb92c9e257fd18fdc500ee5c01b1eb53e
adds support to use neutron as a source of data for
the inventory in addition to the heat stack.

Discovery order will be neutron then heat.

Partial-Implements: blueprint network-data-v2-ports
Change-Id: Ia00ef50d6272fd1b3dd6dbbf7ff6edd2472d7dd0