Inventory from neutron resources
It will first set up the undercloud in the inventory,
then add overcloud data based on disovery of data in
neutron and heat stack if available. Overcloud
inventory data from neutron is added first, then data
from the stack is added.

With the pervasive 'setdefault' usage all values in
the inventory will reflect the value first discovered.
For example, a changed IP address in neutron will be
reflected in the inventory even tough the overcloud
heat stack is not yet aware of the change.

Partial-Implements: blueprint network-data-v2-ports
Change-Id: I9e1851aeb92c9e257fd18fdc500ee5c01b1eb53e