[stable] Raise proper error for None network config
When using custom network configs, if the 'outputs' section is
not correct in the template, there won't be any error in stack
create/update. HostnameNetworkConfigMap would have configs as
null i.e {HOSTNAME: null,} and we'll look for None software
configs using heat api resulting in a traceback not easy to
understand. Let's add a check to notify the user with a proper
error.

This is not required in master and victoria branches, as we've
moved away from using heat software configs for network config.

Closes-Bug: #1907746
Change-Id: If79f7051fec4b4f648c4acf9d62474028e0831bd