Revert "Revert "Fix a couple of Swift healthchecks""
This reverts commit 750b562521b507aab555862f01b74e56f8119861.

Depending on the used procps version, the output of "ps -e" is width
limited and thus not matching the quite long process names. Using
"ps -ef" fixes this, along with a small hack in the piped grep to filter
out grep itself (pgrep itself is not working for this).

Removes swift-*-replicator because it does not seem to be used
anymore in any of the container build configurations.

This change removes the renames from the previous commit because they
are backwards incompatible depending on how the containers are built. We
use healthcheck from the openstack-tripleo-common-containers-base which
is a package. If this package does not have the new file names (or old
filenames) and the containers are built with a different version of
tripleo-common, the container builds fail.  So let's skip renaming of
files so that we can deal with that as a separate issue as a whole.

Change-Id: I3e58de46835307d4332b34d22d2895e6382292dd
(cherry picked from commit 21fd04f038e1a90223e156082f780501846fde3c)